+++ src/main/python/transformation/death_to_condition_occurrence.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import List, TYPE_CHECKING
 import pandas as pd
 
-from ..util.code_cleanup import add_dot_to_icdx_code
+from ..util import add_dot_to_icdx_code
 
 if TYPE_CHECKING:
     from src.main.python.wrapper import Wrapper
